+++ main/core/Edge.java
@@ -4,5 +4,5 @@ public interface Edge   {
 	public Node getSource();
 	public Node getDestination();
 	public String toString();
-	public int getEdgeWeight();
+	public double getEdgeWeight();
 	}
